extract Effect Two
